fix(#77): guard darwin field test; fix mcp/transport unreachable code
Two CI regressions landed after PR #166 merged machine auto-detection:

1. TestMachineProfiler_ScanPopulatesFields ran on ubuntu where sw_vers
   and machdep.cpu.brand_string are unavailable, so every field assertion
   failed. Moved to machine_darwin_test.go with //go:build darwin.

2. internal/mcp/transport.go:200 had unreachable code: the flusher
   reference sat after an infinite loop. Restructured handleSSE to
   implement the keep-alive ticker the comment described, which uses
   flusher.Flush() in the ticker case and eliminates the dead code.
